Add fees to EtherSwap
Fees are proportional to the swap value defined in per million.
They are reimbursed upon refund. The swap value is added as an
argument of the commit, otherwise it would be derived from the
msg.value and would make not nice rounded swap values. Also we would
to store the paid fee value in the swap struct to reimburse it on
refund as rounding make it impossible to figure the msg.value back
from swap value.