Add various improvements
- The reveal is now more robust.
If the user submits a transaction, but doesn’t wait on the page for whatever reason(page crashes, internet connection stops etc) He could refresh the page and we’ll check whether there is "open" commitment for his account and offer him to resume it.
If the user returns to the page after the swap expire time, then we show nothing. We would need a history page where he can view his previous commitments and eventually submit a refund tx.
- How it works
Now on the right of the page we display "How it works" steps. The steps are synchronised with the current state the user is in. (connected wallet, added a value for the swap etc)