Enforce swap start only on supported chains
I’ve added an env variable for the chains we support a swap start. The user is now not able to start a swap from the UI on a different chain.
In addition to this we output some warnings in the UI telling the user to switch the chain or to connect his wallet to start.

I’ve also switched the default network for the dropdown to be the first network defined in NEXT_PUBLIC_ENFORCE_SWAP_ON_CHAINS variable