Add a Max amount button to dump all
Instead of typing the total balance amount one could use the max button to dump all his ETH.
The max is "max - 0.3" - we leave 0.1 for the current transaction cost, 0.1 ETH for eventual gas cost for a refund tx and 0.1 for the tx cost of refund.

This means that if a swap is matched - then there are 0.3 ETH left, but better like that, that to lock all the ETH in the contract and then not being able to withdraw it.