Implement refund & look at expire times
This commit implements a refund option in the swap sale flow. If the swap expires the user will be presented with an option to get his money out of the contract.

The flow also changed a little - now we ask the user first to switch to a different chain and then he sees the target chain swap that we found. We check the times on his swap and on the target swap. If the target swap time left is less than 20 sec we prevent the user from revealing the secrets as there is not enough time for his tx to go through. (unless he sets a mega high tx gas, but he won’t have the time for this in the metamask UI…)