Allow contract compile without specifying a .env
Currently the default _accounts require a seed, which is being provided as an environement variable.
When cloning the repo and just compiling it’s not necessary to provide a private key. If no PRIVKEY_MNEMONIC is provided I’m setting it to the default value specified in the hardhat docs:
https://hardhat.org/hardhat-network/docs/reference