Fix - T&C doesn’t show & hide max button
For now hide the max button when the user has less tahn 0.03 eth. There must be a better way to calculate how much one should leave in the account.

For T&C we basically use a hack. In the exported version every file has .html extension. If we don’t use this when we open the page in new tab it’s not being found. Nextjs deals with pages by loading them with js, but we would like to display the terms and conditions on a new page so the user won’t lose the filled out form. (we could do a modal at one point)